Break down config classes, add option to store cache next  to app exe
